backend/pyproject.toml
[project]
name = "helpiq-backend"
version = "0.1.0"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "openai>=1.50.0",
  "openai-agents>=0.2.0",
  "openai-chatkit>=1.0.0",
  "supabase>=2.6.0",
  "python-dotenv",
]
